added basic login
