+++ app/__init__.py
@@ -1,0 +1,33 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+import os
+NAMESPACE = 'inv'
+
+app = Flask(__name__)
+
+app.config['SECRET_KEY'] = '123456790'
+app.config['DATABASE_FILE'] = 's0inv.sqlite'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + app.config['DATABASE_FILE']
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = True
+app.config['SECURITY_REGISTERABLE'] = True
+
+db = SQLAlchemy(app)
+
+# def build_sample_db():
+#     db.drop_all()
+#     db.create_all()
+#
+# app_dir = os.path.realpath(os.path.dirname(__file__))
+# database_path = os.path.join(app_dir, app.config['DATABASE_FILE'])
+# if not os.path.exists(database_path):
+#     build_sample_db()
+
+from flask_admin import Admin
+admin = Admin(app, name='s0inv', template_mode='bootstrap3')
+
+# Setup Flask-Security
+from flask_security import Security, SQLAlchemyUserDatastore, current_user
+from app import models, views, modelviews
+user_datastore = SQLAlchemyUserDatastore(db, models.User, models.Role)
+security = Security(app, user_datastore)
+modelviews.__init__()
