+++ app/inventory_labels.py
